Add automatic custom workflow generation and installation.
This will help a lot of users since as a general rule of UX, the
lesser the number of steps, the better.

After the changes, when the user creates a workflow using the
`createworkflow` script, the `template.workflow` file is duplicated
and the resultant file is renamed. However this renaming doesn't
actually change the name of the shortcut shown in services. It only
changes the name of the workflow. For changing the name completely,
data inside the `Info.plist` present in the Contents directory of
the duplicated workflow is changed in order to reflect the name of
the workflow the user desires (the name they input along with the
command).

The data inside `document.wkflow` in the same directory as above is
also edited in order to reflect the new workflow name (in order for
it to correctly change the `current_workflow.txt` file).