compatibility: Remove dependency of my machines path.
Changed all instances where `/Users/isham/` was used in order to make
the script available for use to everyone. They now refer to generic
variables which apple uses to point to different files and
folders. Due to this change there were a few complications
introduced where since applescript uses HFS+ filesystem in order to
manipulate files and POSIX is the filesystem I'm comfortable with, I
did not know how to deal with errors about permissions occuring when
I tried to write to files. Hence, for writing purposes exclusively,
POSIX system is still followed.