createworkflow: Add default_browser file in `WorkflowManagement` folder.
I have switched to Firefox recently, and the found the Workflow
Manager lacking in support of that browser. This commit is the
beginning of a series of commits which adds the support.

The flow for this is basically going to be that the user sets a
default browser they use (the default default_browser being Google
Chrome), and the scripts use that config to save tabs from the
browser.