core scripts: Add `Firefox` browser support.
Continuing the parent commit, this commit aims to add tabs saving
and reopening feature to Firefox as well.

`Save All Tabs` -
Firefox doesn't have applescript support to get the urls of open
tabs. This causes a certain level of irritation (they have been
'trying' to add this for 10 years now), but it isn't something that
can't be overcome. So to do that we extract the data from a file
Firefox uses to 'recover' open their previously opened
tabs (~/Library/Application Support/Firefox/Profiles/*.default-release/sessionstore-backups/recovery.json).

Here too, Firefox decided to make things more difficult by
compressing the files in a proprietory format `jsonlz4` since
Firefox 57+ (which is why we don't support lower versions of Firefox
- sigh). This is the reason why we're now using `python3` and `pip3`
with the dependency used to decompress this file, called `lz4`. The
little python script written in the diffs finds out the list of open
tabs URLs.

`OpenPreviouslyOpenedTabs` -
We now have a new delimiter we also use to separate the contents of
the `saved-url-list` file into a list. This is because of the
previous two paragraphs where we add another method of saving the
files (that is saved by a `return` character separating each, while
previously the fetched urls from Chrome saved using a `linefeed`
character).